+++ orchestration/src/test/java/com/sap/ai/sdk/orchestration/OrchestrationModuleConfigTest.java
@@ -51,7 +51,7 @@ class OrchestrationModuleConfigTest {
 
   @Test
   void testStackingInputAndOutputFilter() {
-    final var config = new OrchestrationModuleConfig().withLlmConfig(GPT_4O);
+    final var config = new OrchestrationModuleConfig(GPT_4O);
 
     final var filter =
         new AzureContentFilter()
@@ -73,7 +73,7 @@ class OrchestrationModuleConfigTest {
   @Test
   void testThrowOnEmptyFilterConfig() {
 
-    final var config = new OrchestrationModuleConfig().withLlmConfig(GPT_4O);
+    final var config = new OrchestrationModuleConfig(GPT_4O);
 
     assertThatThrownBy(() -> config.withInputFiltering(new AzureContentFilter()))
         .isInstanceOf(IllegalArgumentException.class)
@@ -90,8 +90,7 @@ class OrchestrationModuleConfigTest {
             .withEntities(DPIEntities.ADDRESS)
             .withMaskGroundingInput(true)
             .withAllowList(List.of("Alice"));
-    var config =
-        new OrchestrationModuleConfig().withLlmConfig(GPT_4O).withMaskingConfig(maskingConfig);
+    var config = new OrchestrationModuleConfig(GPT_4O).withMaskingConfig(maskingConfig);
 
     assertThat(config.getMaskingConfig()).isNotNull();
     assertThat(((MaskingModuleConfigProviders) config.getMaskingConfig()).getProviders())
@@ -118,7 +117,7 @@ class OrchestrationModuleConfigTest {
         DpiMasking.anonymization()
             .withRegex("\\d{3}-\\d{2}-\\d{4}", "***-**-****")
             .withRegex("\\d{2}-\\d{2}-\\d{5}", "**-**-*****");
-    var config = new OrchestrationModuleConfig().withLlmConfig(GPT_4O).withMaskingConfig(masking);
+    var config = new OrchestrationModuleConfig(GPT_4O).withMaskingConfig(masking);
     assertThat(config.getMaskingConfig()).isNotNull();
     assertThat(((MaskingModuleConfigProviders) config.getMaskingConfig()).getProviders())
         .hasSize(1);
@@ -208,8 +207,7 @@ class OrchestrationModuleConfigTest {
         TranslationConfig.translateOutputTo("de-DE").withSourceLanguage("en-US");
 
     final var config =
-        new OrchestrationModuleConfig()
-            .withLlmConfig(GPT_4O)
+        new OrchestrationModuleConfig(GPT_4O)
             .withInputTranslationConfig(inputTranslation)
             .withOutputTranslationConfig(outputTranslation);
 
@@ -272,7 +270,7 @@ class OrchestrationModuleConfigTest {
     Map<String, Object> params = Map.of("foo", "bar");
     String version = "2024-05-13";
     OrchestrationAiModel aiModel = GPT_4O.withParams(params).withVersion(version);
-    var config = new OrchestrationModuleConfig().withLlmConfig(aiModel);
+    var config = new OrchestrationModuleConfig(aiModel);
 
     assertThat(config.getLlmConfig()).isNotNull();
     assertThat(config.getLlmConfig().getName()).isEqualTo(GPT_4O.getName());
@@ -288,8 +286,7 @@ class OrchestrationModuleConfigTest {
   @Test
   void testGroundingConfig() {
     var groundingConfig = Grounding.create();
-    var config =
-        new OrchestrationModuleConfig().withLlmConfig(GPT_4O).withGrounding(groundingConfig);
+    var config = new OrchestrationModuleConfig(GPT_4O).withGrounding(groundingConfig);
 
     assertThat(config.getGroundingConfig()).isNotNull();
     assertThat(config.getGroundingConfig().getType()).isEqualTo(DOCUMENT_GROUNDING_SERVICE);
@@ -311,8 +308,7 @@ class OrchestrationModuleConfigTest {
     var filter1 = DocumentGroundingFilter.create().dataRepositoryType(VECTOR).id("123");
     var filter2 = DocumentGroundingFilter.create().dataRepositoryType(VECTOR).id("234");
     var groundingConfig = Grounding.create().filters(filter1, filter2);
-    var config =
-        new OrchestrationModuleConfig().withLlmConfig(GPT_4O).withGrounding(groundingConfig);
+    var config = new OrchestrationModuleConfig(GPT_4O).withGrounding(groundingConfig);
 
     assertThat(config.getGroundingConfig()).isNotNull();
     var configConfig = config.getGroundingConfig().getConfig();
@@ -335,7 +331,7 @@ class OrchestrationModuleConfigTest {
   void testResponseFormatSchema() {
     var schema = ResponseJsonSchema.fromType(TestClassForSchemaGeneration.class);
     var config =
-        new OrchestrationModuleConfig()
+        new OrchestrationModuleConfig(GPT_4O)
             .withTemplateConfig(TemplateConfig.create().withJsonSchemaResponse(schema));
     assertThat(((Template) config.getTemplateConfig())).isNotNull();
     assertThat(
@@ -348,7 +344,7 @@ class OrchestrationModuleConfigTest {
   @Test
   void testResponseFormatObject() {
     var config =
-        new OrchestrationModuleConfig()
+        new OrchestrationModuleConfig(GPT_4O)
             .withTemplateConfig(TemplateConfig.create().withJsonResponse());
     assertThat(((Template) config.getTemplateConfig())).isNotNull();
     assertThat(
@@ -361,7 +357,7 @@ class OrchestrationModuleConfigTest {
   void testResponseFormatOverwrittenByNewTemplateRef() {
     var schema = ResponseJsonSchema.fromType(TestClassForSchemaGeneration.class);
     var config =
-        new OrchestrationModuleConfig()
+        new OrchestrationModuleConfig(GPT_4O)
             .withTemplateConfig(TemplateConfig.create().withJsonSchemaResponse(schema));
     assertThat(((Template) config.getTemplateConfig())).isNotNull();
     assertThat(